+++ src/hooks/useHabitRow.tsx
@@ -1,0 +1,39 @@
+import { useDisclosure } from '@chakra-ui/react';
+import { useCallback } from 'react';
+import { HabitFormProps } from '../components/HabitDetailForm';
+import { Habit } from '../service/HabitManager';
+import { useHabitActionContext } from './context/useHabitContext';
+import { usePeriodContext } from './context/usePeriodContext';
+
+export const useHabitRow = (habitId: Habit['id']) => {
+  const { year, month, lastDay } = usePeriodContext();
+  const { updateHabitDetail, deleteHabit } = useHabitActionContext();
+  const { isOpen, onOpen, onClose } = useDisclosure();
+
+  const submitHandler: HabitFormProps['submitHandler'] = useCallback(
+    (habitDetail) => {
+      updateHabitDetail({
+        ...habitDetail,
+        year,
+        month,
+        id: habitId,
+      });
+      onClose();
+    },
+    [habitId, month, onClose, updateHabitDetail, year],
+  );
+
+  const deleteHandler = useCallback(() => {
+    deleteHabit(habitId);
+    onClose();
+  }, [deleteHabit, habitId, onClose]);
+
+  return {
+    lastDay,
+    isOpen,
+    onOpen,
+    onClose,
+    submitHandler,
+    deleteHandler,
+  };
+};
